+++ ProjectSetupV2/Controllers/JobStatusController.cs
@@ -25,7 +25,7 @@ namespace ProjectSetupV2.Controllers
         }
 
         // GET: JobStatus/Details/5
-        public async Task<IActionResult> Details(long? id)
+        public async Task<IActionResult> Details(int? id)
         {
             if (id == null)
             {
@@ -65,7 +65,7 @@ namespace ProjectSetupV2.Controllers
         }
 
         // GET: JobStatus/Edit/5
-        public async Task<IActionResult> Edit(long? id)
+        public async Task<IActionResult> Edit(int? id)
         {
             if (id == null)
             {
@@ -85,7 +85,7 @@ namespace ProjectSetupV2.Controllers
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(long id, [Bind("Id,Status")] JobStatus jobStatus)
+        public async Task<IActionResult> Edit(int id, [Bind("Id,Status")] JobStatus jobStatus)
         {
             if (id != jobStatus.Id)
             {
@@ -116,7 +116,7 @@ namespace ProjectSetupV2.Controllers
         }
 
         // GET: JobStatus/Delete/5
-        public async Task<IActionResult> Delete(long? id)
+        public async Task<IActionResult> Delete(int? id)
         {
             if (id == null)
             {
@@ -136,7 +136,7 @@ namespace ProjectSetupV2.Controllers
         // POST: JobStatus/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> DeleteConfirmed(long id)
+        public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var jobStatus = await _context.JobStatus.FindAsync(id);
             _context.JobStatus.Remove(jobStatus);
@@ -144,7 +144,7 @@ namespace ProjectSetupV2.Controllers
             return RedirectToAction(nameof(Index));
         }
 
-        private bool JobStatusExists(long id)
+        private bool JobStatusExists(int id)
         {
             return _context.JobStatus.Any(e => e.Id == id);
         }
